+++ packages/scenes/src/querying/SceneQueryRunner.ts
@@ -428,6 +428,12 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
     this._timeSub = undefined;
     this._timeSubRange = undefined;
 
+    // Reset data to its initial state unless it reached a terminal state (Done or Error).
+    // Incomplete data is discarded so reactivation starts with a clean state.
+    if (this.state.data && ![LoadingState.Done, LoadingState.Error].includes(this.state.data.state)) {
+      this.setState({ data: undefined });
+    }
+
     this._drilldownDependenciesManager.cleanup();
   }
 
@@ -469,7 +475,9 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
 
     this._timeSubRange = timeRange;
     this._timeSub = timeRange.subscribeToState(() => {
-      this.runWithTimeRange(timeRange);
+      // setTimeout to let SceneVariableSet also respond to time range change
+      // So that variables that depend on time range have time to switch to loading state
+      setTimeout(() => this.runWithTimeRange(timeRange), 0);
     });
   }